+++ validator/client/BUILD.bazel
@@ -31,7 +31,6 @@ go_library(
         "//shared/timeutils:go_default_library",
         "//validator/accounts/wallet:go_default_library",
         "//validator/db:go_default_library",
-        "//validator/db/kv:go_default_library",
         "//validator/keymanager:go_default_library",
         "//validator/keymanager/imported:go_default_library",
         "//validator/slashing-protection:go_default_library",
@@ -90,7 +89,6 @@ go_test(
         "//shared/timeutils:go_default_library",
         "//validator/db/kv:go_default_library",
         "//validator/db/testing:go_default_library",
-        "//validator/slashing-protection:go_default_library",
         "@com_github_gogo_protobuf//types:go_default_library",
         "@com_github_golang_mock//gomock:go_default_library",
         "@com_github_hashicorp_golang_lru//:go_default_library",
