+++ validator/slashing-protection/local_attester_protect.go
