+++ validator/slashing-protection/local_proposer_protect.go
